add markdown preview to summary and base for learnlist
